+++ src/components/SearchResultsList/SearchResultsList.test.js
@@ -16,12 +16,12 @@ import {
 
 const mockToggleFilterPane = jest.fn();
 const mockSetSelectedAuthorityRecordContext = jest.fn();
-const mockHistoryReplace = jest.fn();
+const mockHistoryPush = jest.fn();
 
 jest.mock('react-router', () => ({
   ...jest.requireActual('react-router'),
   useHistory: () => ({
-    replace: mockHistoryReplace,
+    push: mockHistoryPush,
   }),
   useLocation: jest.fn().mockReturnValue({ search: '' }),
   useRouteMatch: jest.fn().mockReturnValue({ path: '' }),
@@ -194,8 +194,7 @@ describe('Given SearchResultsList', () => {
         }],
         totalResults: 1,
       });
-
-      expect(mockHistoryReplace).toHaveBeenCalledWith(
+      expect(mockHistoryPush).toHaveBeenCalledWith(
         '/authorities/cbc03a36-2870-4184-9777-0c44d07edfe4?authRefType=Authorized&headingRef=Springfield%20%28Colo.%29',
       );
     });